+++ Sources/SwiftUIKit/Views/TextFieldClearButton.swift
@@ -1,0 +1,34 @@
+//
+//  TextFieldClearButton.swift
+//  SwiftUIKit
+//
+//  Created by Daniel Saidi on 2020-12-18.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+public struct TextFieldClearButton: ViewModifier {
+    @Binding var text: String
+
+    public init(text: Binding<String>) {
+        self._text = text
+    }
+
+    public func body(content: Content) -> some View {
+        HStack {
+            content
+            Spacer()
+            Image(systemName: "multiply.circle.fill")
+                .foregroundColor(.secondary)
+                .opacity(text == "" ? 0 : 1)
+                .onTapGesture { self.text = "" }
+        }
+    }
+}
+
+struct TextFieldClearButton_Previews: PreviewProvider {
+    static var previews: some View {
+        TextField("Enter text", text: .constant("Text")).modifier(TextFieldClearButton(text: .constant("Text")))
+    }
+}